refactor: переименование ProjectPage в ProjectLayout
