+++ src/components/project-layout/project-layout.module.css
@@ -1,4 +1,4 @@
-.page {
+.layout {
   position: relative;
   display: flex;
   flex-direction: column;
