--- src/components/project-page/index.ts
@@ -1,1 +1,0 @@
-export * from './project-page';
