--- src/components/project-page/projec-page-section/index.ts
@@ -1,1 +1,0 @@
-export * from './project-page-section';
